made app worked integrated frontend with backend and authenticated;

+++ notes-api/routes/note.api.js
@@ -13,6 +13,7 @@ router.use('/', function (req, res, next) {
     } else {
         firebaseService.validateToken(req.get('X-Auth-Token')).then(function (user) {
             req.userEmail = user.email;
+            
             next();
         }, function (error) {
             res.status(401).send(error);
@@ -22,22 +23,23 @@ router.use('/', function (req, res, next) {
 
 router.get('/by-id/:id', function (req, res, next) {
     var id = req.params.id;
+    var userEmail = req.userEmail;
     
     NoteService.findById(id)
         .then(function (doc) {
-            console.log(doc);
-
-            res.status(200).json(doc);
+            if (doc.userEmail !== userEmail) {
+                res.status(401).send('Você não tem acesso a notas de outro usuário.');
+            } else {
+                res.status(200).json(doc);    
+            }
         }, function (error) {
             res.status(404).send(error);
         });
 });
 
 router.get('/all', function (req, res, next) {
-    NoteService.findAll()
+    NoteService.findByUserEmail(req.userEmail)
         .then(function (docs) {
-            console.log(docs);
-
             res.status(200).json(docs);
         }, function (error) {
             res.status(404).send(error);
@@ -53,8 +55,6 @@ router.post('/create', function (req, res, next) {
 
     NoteService.create(note)
         .then(function (doc) {
-            console.log(doc);
-
             res.header('Location', '/note/by-id/' + doc._id);
             res.status(201).json(doc);
         }, function (error) {
@@ -63,18 +63,27 @@ router.post('/create', function (req, res, next) {
 });
 
 router.put('/update', function (req, res, next) {
-    var note = new Note({
-        _id: req.body._id,
-        title: req.body.title,
-        description: req.body.description
-    });
-
-    NoteService.update(note)
+    NoteService.findById(req.body._id)
         .then(function (doc) {
-            // the OLD doc
-            console.log(doc);
+            if (doc.userEmail !== req.userEmail) {
+                res.status(401).send('Você não tem acesso a notas de outro usuário.');
+            } else {
+                var note = new Note({
+                    _id: req.body._id,
+                    title: req.body.title,
+                    description: req.body.description
+                });
 
-            res.status(200).json(doc);
+                NoteService.update(note)
+                    .then(function (doc) {
+                        // the OLD doc
+                        console.log(doc);
+                        
+                        res.status(200).json(note);
+                    }, function (error) {
+                        res.status(404).send(error);
+                    });
+            }
         }, function (error) {
             res.status(404).send(error);
         });
@@ -83,12 +92,21 @@ router.put('/update', function (req, res, next) {
 router.delete('/delete/:id', function (req, res, next) {
     var id = req.params.id;
 
-    NoteService.delete(id)
+    NoteService.findById(id)
         .then(function (doc) {
-            // the OLD doc
-            console.log(doc);
+            if (doc.userEmail !== req.userEmail) {
+                res.status(401).send('Você não tem acesso a notas de outro usuário.');
+            } else {
+                NoteService.delete(id)
+                    .then(function (doc) {
+                        // the OLD doc
+                        console.log(doc);
 
-            res.status(200).json(doc);
+                        res.status(200).json(doc);
+                    }, function (error) {
+                        res.status(404).send(error);
+                    });
+            }
         }, function (error) {
             res.status(404).send(error);
         });
